Removed SBP From teams + Added one team
I removed the SBP Tag from all the teams that, although are part of the SBP, they have not yet finished their Milestone 1. The SBP is strict as of when teams can announce they are part of the program, and so we shouldn't have them on our website if they haven't yet completed their first Milestone.